==> Sentimental Analysis/1.preprocessing.ipynb <==
!pip install chardet

import chardet

import pandas as pd

with open("/Users/a1101809/Downloads/lotte.txt",'rt') as f:
    for line in f:
        print(line)
        break
    
        

df_lotte =  pd.read_excel("/Users/a1101809/Downloads/lotte.xlsx")

df_lotte_ = df_lotte['연관어'].value_counts()

df__  =df_lotte_.to_frame()

df__.to_csv('/Users/a1101809/Downloads/environt.csv',encoding='utf-8')

df_lotter_.to_frame().to_csv('/Users/a1101809/Downloads/environt.csv',)

eco_list = ['친환경','환경','탄소배출','종이','포장','배출량','효율',
 '탄소','재활용','절감','에너지','효율','에코','포장','ECO','자원','사용량','에너지',
 '분리배출','기후','태양광','재생','일회용','온실가스','플라스틱','기후','아이스팩','탄소중립','토양','자연환경','절감','전력','리사이클',
 '재생','용기','에코','일회용컵','재활용','감축','배출가스','나무','소재','리필','리사이클','녹색'
]

df_lotte.fillna('',inplace=True)

df_lotte_eco = df_lotte.loc[df_lotte['연관어'].str.contains('test')]

df_lotte_eco

for eco_word in eco_list:
    df_temp = df_lotte.loc[df_lotte['연관어'].str.contains(eco_word)]
    df_lotte_eco= pd.concat([df_lotte_eco,df_temp] )
    print(eco_word)

df_lotte_eco

df_lotte_eco_ = df_lotte_eco.drop_duplicates('문장',keep='first')

df_emart =  pd.read_excel("/Users/a1101809/Downloads/emart.xlsx")

df_emart.fillna('',inplace=True)

df_emart_eco = df_emart.loc[df_emart['연관어'].str.contains('test')]

for eco_word in eco_list:
    df_temp = df_emart.loc[df_emart['연관어'].str.contains(eco_word)]
    df_emart_eco= pd.concat([df_emart_eco,df_temp] )
    print(eco_word)

df_emart_eco_ = df_emart_eco.drop_duplicates('문장',keep='first')

import requests

df_emart_eco_.to_csv('/Users/a1101809/Downloads/emart_eco.tsv',sep='\t',encoding='utf-8')

df_lotte_eco_.to_csv('/Users/a1101809/Downloads/lotte_eco.tsv',sep='\t',encoding='utf-8')

### 중립

### drop not in df_lotte_eco

df_lotte_n = df_lotte.loc[df_lotte.index.isin(df_lotte_eco.index)==False]

df_emart_n = df_emart.loc[df_emart.index.isin(df_emart_eco.index)==False]

### 영업시간은 제외함 환경이라고 볼수있어서

df_lotte_n = df_lotte_n.loc[df_lotte_n['문장'].str.contains('영업시간') == False]

temp=[]

for x in df_lotte_n['문장']:
    x = x.replace('\\n',' ')
    x = x.replace('\n',' ')
    x = x.replace("\\",'')
    x = re.sub(r'\\','',x)
    x = x.replace('.','')
  
    temp.append(x)

df_lotte_n['문장']= temp


df_lotte_n['문장'] = df_lotte_n['문장'].replace('\s+', ' ', regex=True)

df_lotte_n = df_lotte_n.drop_duplicates('문장')

df_lotte_n.index = list(range(len(df_lotte_n)))

len(df_emart_n)

df_emart_n = df_emart_n.loc[df_emart_n['문장'].str.contains('영업시간') == False]

### 간단한 전처리

import re

temp = []

for x in df_emart_n['문장']:
    x = x.replace('\\n',' ')
    x = x.replace('\n',' ')
    x = x.replace("\\",'')
    x = re.sub(r'\\','',x)
    x = x.replace('.','')
  
    temp.append(x)

df_emart_n['문장']= temp


df_emart_n['문장'] = df_emart_n['문장'].replace('\s+', ' ', regex=True)

df_emart_n = df_emart_n.drop_duplicates('문장')

df_emart_n.index = list(range(len(df_emart_n)))

df_lotte_n = df_lotte_n.sort_values(by='문장')
df_emart_n = df_emart_n.sort_values(by='문장')

df_lotte_n.to_csv('lotte_neu.csv',sep='\t',index=False)

df_emart_n.to_csv('emart_neu.csv',sep='\t',index=False)

#### shuffle해서 가져오기

df_lotte_n.sample(4000).to_csv('lotte_neu_4000.csv',sep='\t',index=False)

df_emart_n.sample(4000).to_csv('emart_neu_4000.csv',sep='\t',index=False)

df_lotte_n.sort_values()
