==> Sentimental Analysis/2.positive sentence.ipynb <==
import pandas as pd

df_emart = pd.read_csv('emart_eco.tsv',sep='\t')

df_lotte = pd.read_csv('lotte_eco.tsv',sep='\t')

#### emart

a = 'asdf'

a.replace('a','r')

df_emart['문장'] = df_emart['문장'].map(lambda x : x.replace('\n',' '))

df_emart['문장'][15]

df_emart['문장'] = df_emart['문장'].replace('\s+', ' ', regex=True)

df_emart_ = df_emart[['채널','문장','연관어']]

df_emart__ = df_emart_.drop_duplicates('문장')

df_emart__.to_csv('emart_0504.csv',index=False)

df_emart__['문장'] = df_emart__['문장'].map(lambda x : x.replace('\n',' '))

s = "이마트, 매월 옥외 소등…'어스아워' 캠페인 동참\nWWF 환경 운동 캠페인 어스아"

import re

temp = []

for x in df_emart__['문장']:
    x = x.replace('\\n',' ')
    x = x.replace("\\",'')
    x = re.sub(r'\\','',x)
    x = x.replace('.','')
  
    temp.append(x)

df_emart__['문장'] = temp

df_emart__.to_csv('emart_0504.csv',index=False)

### lotte

df_lotte['문장'] = df_lotte['문장'].map(lambda x : x.replace('\n',' '))

df_lotte['문장'] = df_lotte['문장'].replace('\s+', ' ', regex=True)

df_lotte = df_lotte[['채널','문장','연관어']]

temp = []

for x in df_lotte['문장']:
    x = x.replace('\\n',' ')
    x = x.replace("\\",'')
    x = re.sub(r'\\','',x)
    x = x.replace('.','')
  
    temp.append(x)

df_lotte['문장'] = temp

df_lotte = df_lotte.drop_duplicates('문장')

df_lotte.to_csv('lotte_0504.csv',index=False)

df_lotte
