==> Sentimental Analysis/4.lda_preprocess.ipynb <==
import pandas as pd
import requests
import json

from bs4 import BeautifulSoup

lotte_articles = []

lotte_eco_list= []

http://www.newsway.co.kr/news/view?ud=2023010316000769802
http://www.ftoday.co.kr/news/articleView.html?idxno=248315
http://www.naewoeilbo.com/news/articleView.html?idxno=687332
http://www.naeil.com/news_view/?id_art=447889
http://www.hkbs.co.kr/news/articleView.html?idxno=707836

lotte_string = """
http://news.nate.com/view/20230102n08531?mid=n0100
http://news.nate.com/view/20230110n07447?mid=n0100
http://news.nate.com/view/20230315n25063?mid=n0100
http://news.nate.com/view/20230320n06908?mid=n0100
http://news.nate.com/view/20230206n01364?mid=n0100
http://news.zum.com/articles/80721132
https://n.news.naver.com/mnews/article/032/0003198160?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/008/0004838420?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/025/0003257808?rc=N&ntype=RANKING&sid=001
"""

lotte_string = lotte_string.split('\n')

lotte_string = lotte_string[1:-1]

url = lotte_string[8]

url

headers = {"user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.89 Safari/537.36"}

response = requests.get(url,headers= headers)

content = response.content

soup = BeautifulSoup(content, 'html.parser')

realArtcContents_div = soup.find('div', {'id': 'dic_area'})

string = """
롯데마트는 25년 만에 종이 전단 운영을 전면 중단했다. 롯데마트는 친환경 경영의 일환으로 새해부터 1회용 종이 전단을 대신해 새롭게 개편된 모바일 전단을 운영한다. 이를 통해 롯데마트는 연간 150여 t의 종이 사용을 줄일 수 있게 되었다. 이는 연간 20년생 나무 약 3000그루를 보존하는 것과 같으며, 1만6000여 ㎏의 탄소배출을 저감하는 효과로도 이어진다.

모바일 전단은 롯데마트의 오프라인 전용 애플리케이션인 ‘롯데마트GO’ 또는 매장 내 고지된 QR코드를 스캔해 확인할 수 있으며, 매주 목요일 오전 9시에 해당 주차의 새로운 소식과 행사 내용이 업로드 된다. 또 새해부터 모바일 전단이 종이 전단을 완전히 대체하게 된 만큼 기존에 운영해오던 디지털 전단과는 달리 좀 더 모바일에 특화된 사이즈로 전단 디자인을 개편했다.

롯데마트는 고객과 함께 하는 자원 선순환 활동과 친환경 에너지 사용 확산에도 힘쓰고 있다. 지난해 10월 롯데마트는 환경부 포함 18개 기관이 함께하는 ‘고품질 재활용품 회수·보상 사업’ 추진을 위한 업무 협약을 체결했다. 이를 통해 롯데마트 매장을 고객들의 재활용품 수거 거점으로 지원해 자원 회수 및 보상 공간으로 활용할 계획이다. 지난해 12월에는 1회용컵 보증금 제도 제도화에 따라 제주점에 ‘1회용컵 간이 회수기’를 설치하기도 했다.

김정우 롯데마트 브랜드전략팀 전단 담당자는 “친환경 경영과 ESG가치 실현을 위해 20여 년간 운영하던 1회용 종이 전단 운영을 전면 중단하고 이제는 모바일 전단으로 고객들과 소통하고자 한다”며 “향후 이와 연계된 다양한 프로모션도 검토할 예정”이라고 말했다.
"""

strings = string.split(".")

lotte_articles.append(strings)

#temp = []

# for p_tag in realArtcContents_div.find_all('p'):
#     temp.append(p_tag.text)

#temp

#lotte_articles.append(temp)

lotte_articles

df = pd.DataFrame(lotte_articles)

with open('lda_lotte_eco_list.pickle','wb') as f:
    pickle.dump(lotte_articles,f)

lotte_articles.append(strings)

#### em

requests.get()

emart_string = """
http://news.nate.com/view/20230302n23505?mid=n0100
http://news.nate.com/view/20230110n25792?mid=n0100
http://news.nate.com/view/20230110n30696?mid=n0100
https://n.news.naver.com/mnews/article/448/0000390488?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/421/0006567193?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/011/0004159142?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/243/0000040969?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/016/0002088936?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/001/0013838871?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/469/0000731094?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/277/0005236267?rc=N&ntype=RANKING&sid=001
https://n.news.naver.com/mnews/article/421/0006567193?rc=N&ntype=RANKING&sid=001
"""

emart_articles = []
emart_eco_list= []


emart_string = emart_string.split('\n')
emart_string = emart_string[1:-1]

url = emart_string[2]

url

response = requests.get(url,headers= headers)
content = response.content
soup = BeautifulSoup(content, 'html.parser')

realArtcContents_div = soup.find('div', {'id': 'realArtcContents'})

#realArtcContents_div = soup.find('div', {'id': 'dic_area'})


string = """
이마트가 올 설 선물세트에 친환경 요소를 확대해 고객들에게 좋은 반응을 이끌어내고 있다.

10일 이마트에 따르면 이마트는 축수산 선물세트에 '스티로폼 제로' 전략을 본격 도입했다. 냉장 축산 선물세트 중 약 40%인 15개 품목, 수산 선물세트 중 약 20%인 7개 품목 포장재를 스티로폼을 사용하지 않고 종이 재질로 바꿨다.

선물세트에서 스티로폼을 가장 많이 사용하는 분야는 보냉이 필요한 축산, 수산 선물세트다. 이마트는 절감 효과를 극대화하기 위해 조선호텔 냉장 한우 세트 전품목(10개 상품) 및 피코크 횡성축협 한우 구이 세트 4개 상품, 프리미엄 굴비 세트 등 작년 추석 가장 많이 팔렸던 인기 선물세트에 스티로폼 제로 전략을 도입했다.

스티로폼 제로전략을 위해 포장 재질에 대한 테스트도 진행했다. 가장 중요한 보냉 기능을 위해 포장 종이 두께를 더 두껍게했으며 외부 충격에 잘 견딜 수 있도록 골격을 세밀하게 짰다. 보냉 테스트 결과 스티로폼 보냉과 같은 수준의 효과를 얻은 것으로 나타났다.

이마트는 올 설 선물세트 판매기간 동안 축·수산 선물세트 스티로폼 제로 전략을 통해 약 1.5톤의 스티로폼을 줄일 것으로 예상했다.

아이스팩 역시 친환경 재질로 변경했다. 올 설 축산세트 전품목에 들어가는 아이스팩 성분을 환경파괴가 심했던 고분자폴리머에서 물·전분으로 전환했다. 아이스팩 포장지 역시 단일재질에 분리수거가 가능한 저밀도 폴리에틸렌으로 변경했다.

포장재 내부도 분리배출 방법 등을 표시한 친환경 디자인을 적용했으며 내부 종이도 콩기름 인쇄 용지를 사용했다.

과일세트에도 친환경 포장 방법을 적용했다. 저탄소 한라봉세트를 포함한 이마트 자연주의 만감류 세트 4종을 부직포가 아닌 실종이가방에 담았다. 실종이가방은 실과 종이를 합쳐서 만든 포장재로 분리수거가 불가능한 부직포 가방의 대용품이다. 일반 종이가방보다 내구성이 뛰어나며 친환경 소재로 종이로 분리수거가 가능하다.

이마트가 인기 선물세트에 친환경 전략을 도입한 이유는 친환경 등 가치소비가 주요 소비 트렌트로 자리 잡으면서 고객의 의견을 반영한 상품이 큰 인기를 끌고 있기 때문이다.

가치소비란 소비자 자신이 가치를 부여하거나 지향하는 바에 따라 만족도 높은 상품을 구매하는 형태로, MZ세대부터 시작해 다양한 연령층에게 퍼지며 인기를 얻고 있다.

이마트 자체 농산물 저장센터인 후레쉬센터는 올해 친환경 농산물 취급자 인증을 획득하기도 했다.

친환경 포장재로 변경한 축산세트 역시 매출이 큰 폭으로 상승했다.

이마트가 올해 설 선물세트 사전예약기간인 지난해 12월1일부터 지난 8일까지 매출을 분석한 결과 친환경 포장지로 변경한 축산세트가 작년 동기간 대비 22.4% 신장했다. 축산 전체 선물세트 매출 신장(18%)을 넘어선 것.

피코크 횡성축협 한우 등심세트는 60.5% 신장했으며 피코크 프라임 구이세트는 147.5%라는 고신장을 기록하기도 했다.

김동민 이마트 신선 담당은 "건강한 삶을 지향하고 환경을 우선하는 '가치소비'가 확대되면서 이마트 선물세트에도 친환경 포장재 도입을 확대했다"며 "앞으로도 이마트는 환경적, 사회적 가치를 중시하는 가치소비 활동에 동참할 것"라고 말했다.
"""
strings = string.split(".")


emart_articles.append(strings)

len(emart_articles)

with open('lda_emart_eco_list.pickle','wb') as f:
    pickle.dump(emart_articles,f)

lda_emart = pd.DataFrame(emart_articles)

result = []
for article in emart_articles:
    temp = ''
    for sentence in article:
        temp += sentence
    
    result.append(temp)
    
    

lda_emart = pd.DataFrame(result)

lda_emart[1] = list(range(12))

lda_emart.columns = ['document','document_no']

### lotte

result = []
for article in lotte_articles:
    temp = ''
    for sentence in article:
        temp += sentence
    
    result.append(temp)

lda_lotte = pd.DataFrame(result)

lda_lotte[1] = list(range(11))

lda_lotte.columns = ['document','document_no']

df_lotte_neu  = pd.read_csv('lotte_neu.csv',sep='\t')

df_lotte_neu = df_lotte_neu.loc[df_lotte_neu['채널']=='언론']

df_lotte_neu=df_lotte_neu.sort_values(by=['문장번호'])

df_lotte_neu_group = df_lotte_neu.groupby('문장번호').count()

df_lotte_neu_group = df_lotte_neu_group.reset_index()

df_lotte_neu_group = df_lotte_neu_group[['문장번호','문서번호']]

df_lotte_neu_group.columns = ['문장번호','문장수']

df_lotte_neu_ = df_lotte_neu.merge(df_lotte_neu_group,how='left',on='문장번호')

df_lotte_neu_ = df_lotte_neu_.sort_values(by=['문장수','문장번호'],ascending=False)

df_lotte_neu_14 = df_lotte_neu_.loc[df_lotte_neu_['문장수']<14]

result = []
temp = ''
previous = ''
for i , j in zip ( df_lotte_neu_14['문장번호'] , df_lotte_neu_14['문장']):
    if i != previous:
        result.append(temp)
        temp = j
        previous = i
    else:
        temp += j 
        previous = i 

    
    if len(result) == 51:
        break
        
    

len(result)

result.pop(0)

### 이마트

df_emart_neu  = pd.read_csv('emart_neu.csv',sep='\t')

df_emart_neu = df_emart_neu.loc[df_emart_neu['채널']=='언론']
df_emart_neu=df_emart_neu.sort_values(by=['문장번호'])
df_emart_neu_group = df_emart_neu.groupby('문장번호').count()
df_emart_neu_group = df_emart_neu_group.reset_index()
df_emart_neu_group = df_emart_neu_group[['문장번호','문서번호']]
df_emart_neu_group.columns = ['문장번호','문장수']

df_emart_neu_ = df_emart_neu.merge(df_emart_neu_group,how='left',on='문장번호')

df_emart_neu_ = df_emart_neu_.sort_values(by=['문장수','문장번호'],ascending=False)

df_emart_neu_14 = df_emart_neu_.loc[df_emart_neu_['문장수']<14]

result = []
temp = ''
previous = ''
for i , j in zip ( df_emart_neu_14['문장번호'] , df_emart_neu_14['문장']):
    if i != previous:
        result.append(temp)
        temp = j
        previous = i
    else:
        temp += j 
        previous = i 

    
    if len(result) == 51:
        break

df_emart_neu_14_ = pd.DataFrame(result)

df_emart_neu_14_[1] = list(range(12,62))

df_emart_neu_14_.columns = ['document','document_no']

lda_emart = pd.concat([lda_emart,df_emart_neu_14_])

df_lotte_neu_14_ = pd.DataFrame(result)

df_lotte_neu_14_[1] = list(range(11,61))

df_lotte_neu_14_.columns = ['document','document_no']

lda_lotte = pd.concat([lda_lotte,df_lotte_neu_14_])

lda_emart.to_csv("lda_emart.csv",sep='\t',index=False)

lda_lotte.to_csv("lda_lotte.csv",sep='\t',index=False)

lda_lotte

import pandas as pd

df = pd.read_csv('emart_neu_4000.csv',sep='\t')

df_ = df.loc[df['연관어'].isna()==True]

df_naver = df_.loc[df_['URL'].str.contains('naver')]

df_naver_ = df_naver.loc[df_naver['채널'].str.contains('언론')]

df_naver__ = df_naver_.loc[df_naver_['문장'].str.contains('이마트')]

df_naver__sample = df_naver__.sample(50)

emart_neau_list = list(df_naver__sample['URL'].values)

import pickle

with open('emart_neau_list.pickle','wb') as f:
    pickle.dump(emart_neau_list,f)

### lotte

df = pd.read_csv('lotte_neu_4000.csv',sep='\t')

df_ = df.loc[df['연관어'].isna()==True]

df_lotte = df_.loc[df_['URL'].str.contains('naver')]

df_lotte_ = df_lotte.loc[df_lotte['채널'].str.contains('언론')]

df_lotte__ = df_lotte_.loc[df_lotte_['문장'].str.contains('롯데')]

df_lotte__sample = df_lotte__.sample(50)

lotte_neau_list = list(df_lotte__sample['URL'].values)

with open('lotte_neau_list.pickle','wb') as f:
    pickle.dump(lotte_neau_list,f)
