==> Sentimental Analysis/5.lda.ipynb <==
from konlpy.tag import Mecab
from tqdm import tqdm
import re
import pickle
import csv

import pandas as pd

df = pd.read_csv('lda_lotte.csv',sep='\t',header=None)

df= df[[0]]

df.columns = ['document']

df_1 = pd.read_excel('/Users/a1101809/Downloads/lda_lotte_1.xlsx')

df_lotte = pd.concat([df,df_1])

### emart

df_emart = pd.read_csv('lda_emart.csv',sep='\t')

!pip install mecab-python

mecab= Mecab()


def clean_text(text):
    """
    한글, 영문, 숫자만 남기고 제거한다.
    :param text:
    :return:
    """
    text = text.replace(".", " ").strip()
    text = text.replace("·", " ").strip()
    pattern = '[^ ㄱ-ㅣ가-힣|0-9|a-zA-Z]+'
    text = re.sub(pattern=pattern, repl='', string=text)
    return text


def get_nouns(tokenizer, sentence):
    """
    단어의 길이가 2이상인 일반명사(NNG), 고유명사(NNP), 외국어(SL)만을 반환한다.
    :param tokenizer:
    :param sentence:
    :return:
    """
    tagged = tokenizer.pos(sentence)
    nouns = [s for s, t in tagged if t in ['SL', 'NNG', 'NNP'] and len(s) > 1]
    return nouns


def tokenize(df):
    tokenizer = Mecab(dicpath='C:/mecab/mecab-ko-dic')
    processed_data = []
    for sent in tqdm(df['description']):
        sentence = clean_text(sent.replace('\n', '').strip())
        processed_data.append(get_nouns(tokenizer, sentence))
    return processed_data

def save_processed_data(processed_data):
    """
    토큰 분리한 데이터를 csv로 저장
    :param processed_data:
    :return:
    """
    with open('./data/tokenized_data.csv', 'w', newline='', encoding='utf-8') as f:
        writer = csv.writer(f)
        for data in processed_data:
            writer.writerow(data)


if __name__ == '__main__':
    # 유공자 공훈조서 데이터를 읽어들인다.
    df = pd.read_csv('./data/historical_records.csv')
    processed_data = tokenize(df)   # description 부분을 토크나이징 한다.

    # 토큰 분리한 데이터를 저장
    save_processed_data(processed_data)
